parse-image-POC/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
-e file:///Users/danielwaltrip/all-files/projects/code/my-ml-utils
    # via -r requirements.in
appdirs==1.4.4
    # via ptpython
certifi==2023.5.7
    # via requests
charset-normalizer==3.2.0
    # via requests
contourpy==1.1.0
    # via matplotlib
cycler==0.11.0
    # via matplotlib
filelock==3.12.2
    # via torch
fonttools==4.42.0
    # via matplotlib
idna==3.4
    # via requests
jedi==0.19.0
    # via ptpython
jinja2==3.1.2
    # via torch
kiwisolver==1.4.4
    # via matplotlib
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via jinja2
matplotlib==3.7.2
    # via
    #   -r requirements.in
    #   my-ml-utils
    #   seaborn
    #   ultralytics
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
networkx==3.1
    # via torch
noise==1.2.2
    # via -r requirements.in
numpy==1.25.2
    # via
    #   -r requirements.in
    #   contourpy
    #   matplotlib
    #   my-ml-utils
    #   opencv-python
    #   pandas
    #   scipy
    #   seaborn
    #   torchvision
    #   ultralytics
opencv-python==4.8.0.74
    # via ultralytics
packaging==23.1
    # via matplotlib
pandas==2.0.3
    # via
    #   seaborn
    #   ultralytics
parso==0.8.3
    # via jedi
pillow==10.0.0
    # via
    #   matplotlib
    #   torchvision
    #   ultralytics
prompt-toolkit==3.0.39
    # via ptpython
psutil==5.9.5
    # via ultralytics
ptpython==3.0.23
    # via -r requirements.in
py-cpuinfo==9.0.0
    # via ultralytics
pygments==2.16.1
    # via
    #   ptpython
    #   rich
pyparsing==3.0.9
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2023.3
    # via pandas
pyyaml==6.0.1
    # via ultralytics
requests==2.31.0
    # via
    #   torchvision
    #   ultralytics
rich==13.5.2
    # via -r requirements.in
scipy==1.11.1
    # via ultralytics
seaborn==0.12.2
    # via ultralytics
six==1.16.0
    # via python-dateutil
sympy==1.12
    # via torch
torch==2.0.1
    # via
    #   -r requirements.in
    #   my-ml-utils
    #   torchvision
    #   ultralytics
torchinfo==1.8.0
    # via
    #   -r requirements.in
    #   my-ml-utils
torchvision==0.15.2
    # via
    #   -r requirements.in
    #   my-ml-utils
    #   ultralytics
tqdm==4.65.0
    # via ultralytics
typing-extensions==4.7.1
    # via torch
tzdata==2023.3
    # via pandas
ultralytics==8.0.149
    # via -r requirements.in
urllib3==1.26.16
    # via requests
wcwidth==0.2.6
    # via prompt-toolkit
